Move signature verification into utils file
